django_backend/requirements.txt
django==5.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
python-dotenv==1.0.0
pillow==11.2.1
bcrypt==4.3.0
django-cors-headers==4.7.0
django-tenants==3.8.0
psycopg2-binary==2.9.10